ref(onboarding): Move useScmProviders into components directory
Colocate useScmProviders with the other SCM hooks in the components
subdirectory for consistency.

Refs VDY-19
